Fix header alignment, split .cpp files
