+++ mlir/include/mlir/Dialect/AMDGPU/IR/AMDGPUAttrs.td
@@ -1,4 +1,4 @@
-//===-- AMDGPUAttrs.td - AMDGPU dialect attributes *- tablegen -*---===//
+//===-- AMDGPUAttrs.td - AMDGPU dialect attributes *- tablegen -*----------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
