+++ mlir/include/mlir/Dialect/AMDGPU/IR/AMDGPUDialect.h
@@ -25,7 +25,8 @@
 #include "mlir/Dialect/AMDGPU/IR/AMDGPUDialect.h.inc"
 
 #include "mlir/Dialect/AMDGPU/IR/AMDGPUEnums.h.inc"
-#include "mlir/Dialect/AMDGPU/IR/AMDGPUAttributes.h.inc"
+
+#include "mlir/Dialect/AMDGPU/IR/AMDGPUAttrs.h.inc"
 #include "mlir/Dialect/AMDGPU/IR/AMDGPUTypes.h.inc"
 
 namespace mlir::amdgpu {
@@ -52,7 +53,7 @@ inline void printMNKDimensionList(OpAsmPrinter &printer, Operation *,
 } // namespace mlir::amdgpu
 
 #define GET_ATTRDEF_CLASSES
-#include "mlir/Dialect/AMDGPU/IR/AMDGPUAttributes.h.inc"
+#include "mlir/Dialect/AMDGPU/IR/AMDGPUAttrs.h.inc"
 
 #define GET_TYPEDEF_CLASSES
 #include "mlir/Dialect/AMDGPU/IR/AMDGPUTypes.h.inc"
