+++ mlir/include/mlir/Dialect/AMDGPU/IR/AMDGPUEnums.td
@@ -1,4 +1,4 @@
-//===-- AMDGPUEnums.td - AMDGPU dialect enums *- tablegen -*--------===//
+//===-- AMDGPUEnums.td - AMDGPU dialect enums *- tablegen -*---------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
