+++ mlir/include/mlir/Dialect/AMDGPU/IR/AMDGPUOps.td
@@ -1,4 +1,4 @@
-//===-- AMDGPUOps.td - AMDGPU dialect operations *- tablegen -*----- -===//
+//===-- AMDGPUOps.td - AMDGPU dialect operations *- tablegen -*------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
