+++ mlir/include/mlir/Dialect/AMDGPU/IR/AMDGPUTypes.td
@@ -1,4 +1,4 @@
-//===-- AMDGPUTypes.td - AMDGPU dialect types *- tablegen -*--------===//
+//===-- AMDGPUTypes.td - AMDGPU dialect types *- tablegen -*---------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
