+++ mlir/include/mlir/Dialect/AMDGPU/IR/CMakeLists.txt
@@ -7,6 +7,6 @@ mlir_tablegen(AMDGPUEnums.cpp.inc -gen-enum-defs)
 add_mlir_dialect_tablegen_target(MLIRAMDGPUEnumsGen)
 
 set(LLVM_TARGET_DEFINITIONS AMDGPUAttrs.td)
-mlir_tablegen(AMDGPUAttributes.h.inc -gen-attrdef-decls -attrdefs-dialect=amdgpu)
-mlir_tablegen(AMDGPUAttributes.cpp.inc -gen-attrdef-defs -attrdefs-dialect=amdgpu)
+mlir_tablegen(AMDGPUAttrs.h.inc -gen-attrdef-decls -attrdefs-dialect=amdgpu)
+mlir_tablegen(AMDGPUAttrs.cpp.inc -gen-attrdef-defs -attrdefs-dialect=amdgpu)
 add_mlir_dialect_tablegen_target(MLIRAMDGPUAttributesIncGen)
