+++ mlir/lib/Dialect/AMDGPU/IR/AMDGPUAttrs.cpp
@@ -1,0 +1,26 @@
+//===- AMDGPUAttrs.cpp - MLIR AMDGPU dialect attributes -------------------===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+//
+// This file implements the AMDGPU dialect attributes.
+//
+//===----------------------------------------------------------------------===//
+
+#include "mlir/Dialect/AMDGPU/IR/AMDGPUDialect.h"
+
+#include "mlir/IR/DialectImplementation.h"
+#include "llvm/ADT/TypeSwitch.h"
+
+#define GET_ATTRDEF_CLASSES
+#include "mlir/Dialect/AMDGPU/IR/AMDGPUAttrs.cpp.inc"
+
+void mlir::amdgpu::AMDGPUDialect::registerAttributes() {
+  addAttributes<
+#define GET_ATTRDEF_LIST
+#include "mlir/Dialect/AMDGPU/IR/AMDGPUAttrs.cpp.inc"
+      >();
+}
