+++ mlir/lib/Dialect/AMDGPU/IR/CMakeLists.txt
@@ -1,5 +1,9 @@
 add_mlir_dialect_library(MLIRAMDGPUDialect
+  AMDGPUAttrs.cpp
   AMDGPUDialect.cpp
+  AMDGPUEnums.cpp
+  AMDGPUOps.cpp
+  AMDGPUTypes.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${MLIR_MAIN_INCLUDE_DIR}/mlir/Dialect/AMDGPU
